fix bug in $.fn.fitText
When the text element to be dynamically sized was directly inside a block
element with an automatically determined width, e.g. a div with no matching CSS
width rule or in-line style property, fitText would still use the current width
of the parent as the width to fit the text into. In effect, the text
would shrink with the parent when the window became narrower than the parent
element, but would not grow again because the parent did not grow. The
solution, for now, is to only consider the width of the parent if it is
set in-line. Otherwise just use the window width.